changed admin page to new format
